GPII-3646:  Refactored common database requests
Moved data base requests to the shared folder as a set of utility
functions and modified the deletion script to use them.
